requirements.txt
textual>=0.85.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
pydantic>=2.0.0
networkx>=3.0
python-dotenv>=1.0.0
